ToolWhyItWorked refactor and redirect of translation
