+++ app/components/ToolWhyItWorked/Content.js
@@ -1,0 +1,10 @@
+import ContentBlock from 'components/ContentBlock';
+import styled from 'styled-components';
+export default styled(ContentBlock)`
+  img {
+    width: 100%;
+  }
+  li {
+    margin-bottom: 20px;
+  }
+`;
