+++ app/containers/ToolWhyItWorked/staticText.js
@@ -1,0 +1,6 @@
+export default {
+  header: {
+    id: 'app.components.WhyItWorked.header',
+    defaultMessage: 'Why it Worked',
+  },
+};
